+++ cli/tests/test_git_push.rs
@@ -2170,16 +2170,39 @@ fn test_git_push_sign_on_push() {
             ["new", "-m", "commit which should not be signed 1"],
         )
         .success();
+    test_env.run_jj_in(&workspace_root, ["new"]).success();
     test_env
         .run_jj_in(
             &workspace_root,
-            ["new", "-m", "commit which should not be signed 2"],
+            [
+                "describe",
+                "-m",
+                "commit which should not be signed 2 with different author",
+                "--author=Foo <foo@example.org>",
+            ],
         )
+        .success();
+    test_env.run_jj_in(&workspace_root, ["new"]).success();
+    test_env
+        .run_jj_in(
+            &workspace_root,
+            [
+                "describe",
+                "-m",
+                "commit with different author",
+                "--author=Foo <foo@example.org>",
+            ],
+        )
+        .success();
+    test_env
+        .run_jj_in(&workspace_root, ["new", "-m", "final commit"])
         .success();
     // There should be no signed commits initially
     let output = test_env.run_jj_in(&workspace_root, ["log", "-T", template]);
     insta::assert_snapshot!(output, @r"
-    @  commit which should not be signed 2
+    @  final commit
+    ○  commit with different author
+    ○  commit which should not be signed 2 with different author
     ○  commit which should not be signed 1
     ○  commit to be signed 2
     ○  commit to be signed 1
@@ -2193,7 +2216,8 @@ fn test_git_push_sign_on_push() {
         r#"
     signing.backend = "test"
     signing.key = "impeccable"
-    git.sign-on-push = true
+    git.sign-on-push.enabled = true
+    git.sign-on-push.behavior = "own"
     "#,
     );
     let output = test_env.run_jj_in(&workspace_root, ["git", "push", "--dry-run"]);
@@ -2207,7 +2231,9 @@ fn test_git_push_sign_on_push() {
     // There should be no signed commits after performing a dry run
     let output = test_env.run_jj_in(&workspace_root, ["log", "-T", template]);
     insta::assert_snapshot!(output, @r"
-    @  commit which should not be signed 2
+    @  final commit
+    ○  commit with different author
+    ○  commit which should not be signed 2 with different author
     ○  commit which should not be signed 1
     ○  commit to be signed 2
     ○  commit to be signed 1
@@ -2221,17 +2247,19 @@ fn test_git_push_sign_on_push() {
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
     Updated signatures of 2 commits
-    Rebased 2 descendant commits
+    Rebased 4 descendant commits
     Changes to push to origin:
-      Move forward bookmark bookmark2 from 8476341eb395 to a6259c482040
-    Working copy now at: kmkuslsw b5f47345 (empty) commit which should not be signed 2
-    Parent commit      : kpqxywon 90df08d3 (empty) commit which should not be signed 1
+      Move forward bookmark bookmark2 from 8476341eb395 to 2ac4cdcf94ef
+    Working copy now at: nkmrtpmo a4e979ab (empty) final commit
+    Parent commit      : lylxulpl 2b7318f0 (empty) commit with different author
     [EOF]
     ");
     // Only commits which are being pushed should be signed
     let output = test_env.run_jj_in(&workspace_root, ["log", "-T", template]);
     insta::assert_snapshot!(output, @r"
-    @  commit which should not be signed 2
+    @  final commit
+    ○  commit with different author
+    ○  commit which should not be signed 2 with different author
     ○  commit which should not be signed 1
     ○  commit to be signed 2
     │  Signature: test-display, Status: good, Key: impeccable
@@ -2257,7 +2285,7 @@ fn test_git_push_sign_on_push() {
     );
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
-    Created 1 bookmarks pointing to kpqxywon 90df08d3 bookmark3 | (empty) commit which should not be signed 1
+    Created 1 bookmarks pointing to kpqxywon af823162 bookmark3 | (empty) commit which should not be signed 1
     [EOF]
     ");
     let output = test_env.run_jj_in(
@@ -2266,22 +2294,113 @@ fn test_git_push_sign_on_push() {
     );
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
-    Moved 1 bookmarks to kpqxywon 90df08d3 bookmark2* bookmark3 | (empty) commit which should not be signed 1
+    Moved 1 bookmarks to kpqxywon af823162 bookmark2* bookmark3 | (empty) commit which should not be signed 1
     [EOF]
     ");
     test_env.add_config(r#"revset-aliases."immutable_heads()" = "bookmark3""#);
-    let output = test_env.run_jj_in(&workspace_root, ["git", "push"]);
+    let output = test_env.run_jj_in(&workspace_root, ["git", "push", "-b", "bookmark2"]);
     insta::assert_snapshot!(output, @r"
     ------- stderr -------
-    Warning: Refusing to create new remote bookmark bookmark3@origin
-    Hint: Use --allow-new to push new bookmark. Use --remote to specify the remote to push to.
     Changes to push to origin:
-      Move forward bookmark bookmark2 from a6259c482040 to 90df08d3d612
+      Move forward bookmark bookmark2 from 2ac4cdcf94ef to af8231621fcb
     [EOF]
     ");
     let output = test_env.run_jj_in(&workspace_root, ["log", "-T", template, "-r", "::"]);
     insta::assert_snapshot!(output, @r"
-    @  commit which should not be signed 2
+    @  final commit
+    ○  commit with different author
+    ○  commit which should not be signed 2 with different author
+    ◆  commit which should not be signed 1
+    ◆  commit to be signed 2
+    │  Signature: test-display, Status: good, Key: impeccable
+    ◆  commit to be signed 1
+    │  Signature: test-display, Status: good, Key: impeccable
+    ◆  description 2
+    │ ○  description 1
+    ├─╯
+    ◆
+    [EOF]
+    ");
+
+    // With with `behavior = "own"`, commit with different author should not be
+    // signed
+    let output = test_env.run_jj_in(
+        &workspace_root,
+        [
+            "bookmark",
+            "move",
+            "bookmark2",
+            "--to",
+            "description('commit which should not be signed 2')",
+        ],
+    );
+    insta::assert_snapshot!(output, @r"
+    ------- stderr -------
+    Moved 1 bookmarks to kmkuslsw 4f13590e bookmark2* | (empty) commit which should not be signed 2 with different author
+    [EOF]
+    ");
+    let output = test_env.run_jj_in(&workspace_root, ["git", "push", "-b", "bookmark2"]);
+    insta::assert_snapshot!(output, @r"
+    ------- stderr -------
+    Changes to push to origin:
+      Move forward bookmark bookmark2 from af8231621fcb to 4f13590e2672
+    [EOF]
+    ");
+    let output = test_env.run_jj_in(&workspace_root, ["log", "-T", template, "-r", "::"]);
+    insta::assert_snapshot!(output, @r"
+    @  final commit
+    ○  commit with different author
+    ○  commit which should not be signed 2 with different author
+    ◆  commit which should not be signed 1
+    ◆  commit to be signed 2
+    │  Signature: test-display, Status: good, Key: impeccable
+    ◆  commit to be signed 1
+    │  Signature: test-display, Status: good, Key: impeccable
+    ◆  description 2
+    │ ○  description 1
+    ├─╯
+    ◆
+    [EOF]
+    ");
+
+    // With with `behavior = "force"`, commit with different author is signed
+    test_env.add_config(
+        r#"
+    git.sign-on-push.behavior = "force"
+    "#,
+    );
+    let output = test_env.run_jj_in(
+        &workspace_root,
+        [
+            "bookmark",
+            "move",
+            "bookmark2",
+            "--to",
+            "description('commit with different author')",
+        ],
+    );
+    insta::assert_snapshot!(output, @r"
+    ------- stderr -------
+    Moved 1 bookmarks to lylxulpl 2b7318f0 bookmark2* | (empty) commit with different author
+    [EOF]
+    ");
+    let output = test_env.run_jj_in(&workspace_root, ["git", "push", "-b", "bookmark2"]);
+    insta::assert_snapshot!(output, @r"
+    ------- stderr -------
+    Updated signatures of 1 commits
+    Rebased 1 descendant commits
+    Changes to push to origin:
+      Move forward bookmark bookmark2 from 4f13590e2672 to aaed4b63a38a
+    Working copy now at: nkmrtpmo 10ff3e13 (empty) final commit
+    Parent commit      : lylxulpl aaed4b63 bookmark2 | (empty) commit with different author
+    [EOF]
+    ");
+    let output = test_env.run_jj_in(&workspace_root, ["log", "-T", template, "-r", "::"]);
+    insta::assert_snapshot!(output, @r"
+    @  final commit
+    ○  commit with different author
+    │  Signature: test-display, Status: good, Key: impeccable
+    ○  commit which should not be signed 2 with different author
     ◆  commit which should not be signed 1
     ◆  commit to be signed 2
     │  Signature: test-display, Status: good, Key: impeccable